refactor: modularize admin services
