+++ frontend/src/components/Admin/DataManagement.tsx
@@ -47,7 +47,7 @@ import {
   TabsList,
   TabsTrigger,
 } from '@/design-system/components/Tabs';
-import AdminApiService from '@/services/adminApi';
+import * as AdminApi from '@/services/admin';
 import { toast } from 'sonner';
 
 interface TableInfo {
@@ -80,10 +80,10 @@ const DataManagement: React.FC = () => {
     '1h' | '24h' | '7d'
   >('24h');
   const [schedules, setSchedules] = useState<
-    import('@/services/adminApi').MaintenanceSchedules | null
+    import('@/services/admin').MaintenanceSchedules | null
   >(null);
   const [schedulesDraft, setSchedulesDraft] = useState<
-    import('@/services/adminApi').MaintenanceSchedules | null
+    import('@/services/admin').MaintenanceSchedules | null
   >(null);
   const [savingSchedules, setSavingSchedules] = useState(false);
 
@@ -94,13 +94,13 @@ const DataManagement: React.FC = () => {
 
       const [tables, health, performance, integrity, sched] = await Promise.all(
         [
-          AdminApiService.getTableInformation(),
-          AdminApiService.getDatabaseHealth(),
-          AdminApiService.getDatabasePerformance(10, {
+          AdminApi.getTableInformation(),
+          AdminApi.getDatabaseHealth(),
+          AdminApi.getDatabasePerformance(10, {
             window: performanceWindow,
           }),
-          AdminApiService.checkDataIntegrity(),
-          AdminApiService.getMaintenanceSchedules(),
+          AdminApi.checkDataIntegrity(),
+          AdminApi.getMaintenanceSchedules(),
         ]
       );
 
@@ -132,7 +132,7 @@ const DataManagement: React.FC = () => {
   // Preview cleanup
   const previewCleanup = async () => {
     try {
-      const result = await AdminApiService.cleanupFiles(true);
+      const result = await AdminApi.cleanupFiles(true);
       setCleanupPreview({
         orphaned_files: result.orphaned_files,
         failed_files: result.failed_files,
@@ -149,7 +149,7 @@ const DataManagement: React.FC = () => {
   const executeCleanup = async () => {
     try {
       setCleanupInProgress(true);
-      const result = await AdminApiService.cleanupFiles(false);
+      const result = await AdminApi.cleanupFiles(false);
       toast.success(result.message);
       setShowCleanupDialog(false);
       await loadDataInfo(); // Refresh data
@@ -163,7 +163,7 @@ const DataManagement: React.FC = () => {
   // Database cleanup
   const handleDatabaseCleanup = async (dryRun = true) => {
     try {
-      const result = await AdminApiService.cleanupDatabase(dryRun);
+      const result = await AdminApi.cleanupDatabase(dryRun);
       if (dryRun) {
         toast.success(`Cleanup preview: ${JSON.stringify(result)}`);
       } else {
@@ -213,7 +213,7 @@ const DataManagement: React.FC = () => {
   useEffect(() => {
     (async () => {
       try {
-        const perf = await AdminApiService.getDatabasePerformance(10, {
+        const perf = await AdminApi.getDatabasePerformance(10, {
           window: performanceWindow,
         });
         setPerformanceData(perf);
@@ -435,7 +435,7 @@ const DataManagement: React.FC = () => {
                                     return;
                                   }
                                   const res =
-                                    await AdminApiService.exportDatabase({
+                                    await AdminApi.exportDatabase({
                                       table: table.name,
                                       format: fmt as 'json' | 'csv',
                                     });
@@ -545,7 +545,7 @@ const DataManagement: React.FC = () => {
                         if (!confirm('Generate admin overview report (CSV)?'))
                           return;
                         const data =
-                          await AdminApiService.getAdminOverviewReport('csv');
+                          await AdminApi.getAdminOverviewReport('csv');
                         const blob = new Blob([data], { type: 'text/csv' });
                         const url = URL.createObjectURL(blob);
                         const a = document.createElement('a');
@@ -786,7 +786,7 @@ const DataManagement: React.FC = () => {
                           try {
                             setSavingSchedules(true);
                             const saved =
-                              await AdminApiService.updateMaintenanceSchedules(
+                              await AdminApi.updateMaintenanceSchedules(
                                 schedulesDraft!
                               );
                             setSchedules(saved);
@@ -823,7 +823,7 @@ const DataManagement: React.FC = () => {
                   onClick={async () => {
                     try {
                       if (!confirm('Run database backup now?')) return;
-                      const res = await AdminApiService.backupDatabase();
+                      const res = await AdminApi.backupDatabase();
                       toast.success(`${res.message} (job ${res.job_id})`);
                     } catch {
                       toast.error('Backup failed');
@@ -850,7 +850,7 @@ const DataManagement: React.FC = () => {
                         toast.error('Invalid format');
                         return;
                       }
-                      const res = await AdminApiService.exportDatabase({
+                      const res = await AdminApi.exportDatabase({
                         table,
                         format: format as 'json' | 'csv',
                       });
@@ -871,7 +871,7 @@ const DataManagement: React.FC = () => {
                   onClick={async () => {
                     try {
                       if (!confirm('Rebuild all database indexes now?')) return;
-                      const res = await AdminApiService.reindexDatabase();
+                      const res = await AdminApi.reindexDatabase();
                       toast.success(`${res.message} (job ${res.job_id})`);
                     } catch {
                       toast.error('Reindex failed');
