+++ frontend/src/components/Admin/UserManagement.tsx
@@ -71,7 +71,8 @@ import {
 } from '@/design-system/components/DropdownMenu';
 import { Label } from '@/design-system/components/Label';
 // import { Textarea } from '@/design-system/components/Textarea';
-import AdminApiService, { UserWithRoles } from '@/services/adminApi';
+import * as AdminApi from '@/services/admin';
+import type { UserWithRoles } from '@/services/admin';
 import { toast } from 'sonner';
 
 interface UserManagementProps {
@@ -124,7 +125,7 @@ const UserManagement: React.FC<UserManagementProps> = ({ onUserUpdated }) => {
   const loadUsers = async () => {
     try {
       setLoading(true);
-      const resp = await AdminApiService.listUsers(0, 200, true);
+      const resp = await AdminApi.listUsers(0, 200, true);
       const env = resp as any;
       const list = (env?.items as UserWithRoles[]) || (resp as UserWithRoles[]);
       setUsers(list);
@@ -201,7 +202,7 @@ const UserManagement: React.FC<UserManagementProps> = ({ onUserUpdated }) => {
     }
 
     try {
-      const result = await AdminApiService.bulkUserAction(
+      const result = await AdminApi.bulkUserAction(
         selectedUsers,
         action
       );
@@ -232,7 +233,7 @@ const UserManagement: React.FC<UserManagementProps> = ({ onUserUpdated }) => {
     if (!editingUser) return;
 
     try {
-      await AdminApiService.updateUser(editingUser.id, editUserData);
+      await AdminApi.updateUser(editingUser.id, editUserData);
       toast.success('User updated successfully');
       setShowEditDialog(false);
       setEditingUser(null);
@@ -248,7 +249,7 @@ const UserManagement: React.FC<UserManagementProps> = ({ onUserUpdated }) => {
     if (!userToDelete) return;
 
     try {
-      await AdminApiService.deleteUser(userToDelete.id);
+      await AdminApi.deleteUser(userToDelete.id);
       toast.success('User deactivated successfully');
       setShowDeleteDialog(false);
       setUserToDelete(null);
@@ -263,7 +264,7 @@ const UserManagement: React.FC<UserManagementProps> = ({ onUserUpdated }) => {
   const handleCreateUser = async () => {
     try {
       setCreating(true);
-      await AdminApiService.createUser(newUserData);
+      await AdminApi.createUser(newUserData);
       toast.success('User created successfully');
       setShowAddDialog(false);
       setNewUserData({
@@ -286,7 +287,7 @@ const UserManagement: React.FC<UserManagementProps> = ({ onUserUpdated }) => {
   // Role management
   const handleAssignRole = async (userId: number, role: string) => {
     try {
-      await AdminApiService.assignRole(userId, role);
+      await AdminApi.assignRole(userId, role);
       toast.success(`Role ${role} assigned successfully`);
       await loadUsers();
     } catch (error) {
@@ -296,7 +297,7 @@ const UserManagement: React.FC<UserManagementProps> = ({ onUserUpdated }) => {
 
   const handleRemoveRole = async (userId: number, role: string) => {
     try {
-      await AdminApiService.removeRole(userId, role);
+      await AdminApi.removeRole(userId, role);
       toast.success(`Role ${role} removed successfully`);
       await loadUsers();
     } catch (error) {
@@ -329,8 +330,8 @@ const UserManagement: React.FC<UserManagementProps> = ({ onUserUpdated }) => {
 
     try {
       await Promise.all([
-        ...toAssign.map(r => AdminApiService.assignRole(roleUser.id, r)),
-        ...toRemove.map(r => AdminApiService.removeRole(roleUser.id, r)),
+        ...toAssign.map(r => AdminApi.assignRole(roleUser.id, r)),
+        ...toRemove.map(r => AdminApi.removeRole(roleUser.id, r)),
       ]);
       toast.success('Roles updated');
       setShowRolesDialog(false);
