+++ frontend/src/pages/__tests__/AdminDashboard.overview-maintenance.test.tsx
@@ -3,14 +3,12 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
-import * as AdminApi from '@/services/adminApi';
+import * as AdminApi from '@/services/admin';
 import AdminDashboard from '@/pages/AdminDashboard';
 
-vi.mock('@/services/adminApi');
+vi.mock('@/services/admin');
 
-const mocked = AdminApi as unknown as {
-    default: any;
-};
+const mocked = AdminApi as unknown as Record<string, any>;
 
 const baseStats = {
     users: { total: 1000, active: 800, verified: 700, new_24h: 25 },
@@ -47,7 +45,7 @@ describe('AdminDashboard Overview and Maintenance', () => {
     beforeEach(() => {
         vi.resetAllMocks();
 
-        mocked.default = {
+        Object.assign(mocked, {
             getSystemStats: vi.fn().mockResolvedValue(baseStats),
             getUserActivity: vi.fn().mockResolvedValue([
                 {
@@ -70,7 +68,7 @@ describe('AdminDashboard Overview and Maintenance', () => {
             getAuditLogs: vi.fn().mockResolvedValue({ items: [], total: 0, skip: 0, limit: 100 }),
             cleanupFiles: vi.fn().mockResolvedValue({ message: 'Cleanup done', orphaned_files: 0, failed_files: 0 }),
             clearRateLimits: vi.fn().mockResolvedValue({ message: 'Cleared', cleared_records: 1 }),
-        };
+        });
     });
 
     it('renders healthy overview with no critical alerts', async () => {
@@ -83,7 +81,7 @@ describe('AdminDashboard Overview and Maintenance', () => {
     });
 
     it('shows High CPU Usage alert when cpu_usage > 90', async () => {
-        mocked.default.getSystemMetrics = vi.fn().mockResolvedValue(baseMetricsHighCpu);
+        mocked.getSystemMetrics = vi.fn().mockResolvedValue(baseMetricsHighCpu);
 
         render(<AdminDashboard />);
 
@@ -101,13 +99,11 @@ describe('AdminDashboard Overview and Maintenance', () => {
         // Preview Cleanup
         const previewBtn = await screen.findByRole('button', { name: /preview cleanup/i });
         await userEvent.click(previewBtn);
-        expect(mocked.default.cleanupFiles).toHaveBeenCalledWith(true);
+        expect(mocked.cleanupFiles).toHaveBeenCalledWith(true);
 
         // Run Cleanup
         const runBtn = await screen.findByRole('button', { name: /run cleanup/i });
         await userEvent.click(runBtn);
-        expect(mocked.default.cleanupFiles).toHaveBeenCalledWith(false);
+        expect(mocked.cleanupFiles).toHaveBeenCalledWith(false);
     });
 });
-
-
