+++ frontend/src/pages/__tests__/AdminDashboard.robustness.test.tsx
@@ -2,13 +2,13 @@ import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { render, waitFor } from '@testing-library/react'
 import React from 'react'
 
-import * as AdminApi from '@/services/adminApi'
+import * as AdminApi from '@/services/admin'
 import { toast } from 'sonner'
 import AdminDashboard from '@/pages/AdminDashboard'
 
-vi.mock('@/services/adminApi')
+vi.mock('@/services/admin')
 
-const mocked = AdminApi as unknown as { default: any }
+const mocked = AdminApi as unknown as Record<string, any>
 
 describe('AdminDashboard robustness to API failures', () => {
     beforeEach(() => {
@@ -19,7 +19,7 @@ describe('AdminDashboard robustness to API failures', () => {
     })
 
     it('shows warning toast when some sections fail', async () => {
-        mocked.default = {
+        Object.assign(mocked, {
             getSystemStats: vi.fn().mockResolvedValue({ users: {}, files: {}, financial_data: {}, system: {}, performance: {} }),
             getUserActivity: vi.fn().mockResolvedValue([]),
             getSystemMetrics: vi.fn().mockRejectedValue(new Error('fail')),
@@ -30,14 +30,14 @@ describe('AdminDashboard robustness to API failures', () => {
             getSystemLogs: vi.fn().mockResolvedValue([]),
             getUserPermissions: vi.fn().mockResolvedValue({ permissions: [], roles: [], is_admin: true }),
             getAuditLogs: vi.fn().mockResolvedValue({ items: [], skip: 0, limit: 100, total: 0 }),
-        }
+        })
         render(<AdminDashboard />)
         await waitFor(() => expect((toast.warning as any).mock.calls.length).toBeGreaterThanOrEqual(1))
     })
 
     it('shows error toast when all sections fail', async () => {
         const reject = vi.fn().mockRejectedValue(new Error('down'))
-        mocked.default = {
+        Object.assign(mocked, {
             getSystemStats: reject,
             getUserActivity: reject,
             getSystemMetrics: reject,
@@ -48,10 +48,8 @@ describe('AdminDashboard robustness to API failures', () => {
             getSystemLogs: reject,
             getUserPermissions: reject,
             getAuditLogs: reject,
-        }
+        })
         render(<AdminDashboard />)
         await waitFor(() => expect((toast.error as any).mock.calls.length).toBeGreaterThanOrEqual(1))
     })
 })
-
-
