+++ frontend/src/services/__tests__/adminApi.contract.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
-import AdminApiService from '../adminApi'
+import * as AdminApi from '../admin'
 import apiClient from '../api'
 
 vi.mock('../api', () => {
@@ -22,32 +22,32 @@ const api = apiClient as unknown as {
 
 const ok = (data: any) => Promise.resolve({ data })
 
-describe('AdminApiService contracts', () => {
+describe('AdminApi contracts', () => {
   beforeEach(() => {
     vi.clearAllMocks()
   })
 
   it('getSystemStats GET /admin/stats', async () => {
     api.get.mockResolvedValueOnce(ok({}))
-    await AdminApiService.getSystemStats()
+    await AdminApi.getSystemStats()
     expect(api.get).toHaveBeenCalledWith('/admin/stats')
   })
 
   it('getUserActivity GET /admin/users/activity-list with params', async () => {
     api.get.mockResolvedValueOnce(ok([]))
-    await AdminApiService.getUserActivity(20, true)
+    await AdminApi.getUserActivity(20, true)
     expect(api.get).toHaveBeenCalledWith('/admin/users/activity-list', { params: { limit: 20, active_only: true } })
   })
 
   it('getSecurityAudit maps from/to to from_ts/to_ts', async () => {
     api.get.mockResolvedValueOnce(ok({}))
-    await AdminApiService.getSecurityAudit({ from: '2024-01-01', to: '2024-02-01' })
+    await AdminApi.getSecurityAudit({ from: '2024-01-01', to: '2024-02-01' })
     expect(api.get).toHaveBeenCalledWith('/admin/security/audit', { params: { from_ts: '2024-01-01', to_ts: '2024-02-01' } })
   })
 
   it('getSystemLogs maps options and envelope default', async () => {
     api.get.mockResolvedValueOnce(ok([]))
-    await AdminApiService.getSystemLogs('ERROR', 50, { skip: 10, from: 'a', to: 'b', search: 'q', envelope: true })
+    await AdminApi.getSystemLogs('ERROR', 50, { skip: 10, from: 'a', to: 'b', search: 'q', envelope: true })
     expect(api.get).toHaveBeenCalledWith('/admin/system/logs', {
       params: { level: 'ERROR', limit: 50, skip: 10, from_ts: 'a', to_ts: 'b', search: 'q', envelope: true },
     })
@@ -55,7 +55,7 @@ describe('AdminApiService contracts', () => {
 
   it('getAuditLogs maps filters and envelope', async () => {
     api.get.mockResolvedValueOnce(ok([]))
-    await AdminApiService.getAuditLogs(5, 25, 123, 'LOGIN', { from: 'a', to: 'b', envelope: true })
+    await AdminApi.getAuditLogs(5, 25, 123, 'LOGIN', { from: 'a', to: 'b', envelope: true })
     expect(api.get).toHaveBeenCalledWith('/admin/audit-logs', {
       params: { skip: 5, limit: 25, user_id: 123, action: 'LOGIN', from_ts: 'a', to_ts: 'b', envelope: true },
     })
@@ -63,33 +63,31 @@ describe('AdminApiService contracts', () => {
 
   it('cleanupFiles POST with dry_run param', async () => {
     api.post.mockResolvedValueOnce(ok({}))
-    await AdminApiService.cleanupFiles(false)
+    await AdminApi.cleanupFiles(false)
     expect(api.post).toHaveBeenCalledWith('/admin/files/cleanup', null, { params: { dry_run: false } })
   })
 
   it('clearRateLimits POST no body', async () => {
     api.post.mockResolvedValueOnce(ok({}))
-    await AdminApiService.clearRateLimits()
+    await AdminApi.clearRateLimits()
     expect(api.post).toHaveBeenCalledWith('/admin/rate-limits/clear')
   })
 
   it('listUsers GET with pagination and envelope', async () => {
     api.get.mockResolvedValueOnce(ok({}))
-    await AdminApiService.listUsers(2, 10, true)
+    await AdminApi.listUsers(2, 10, true)
     expect(api.get).toHaveBeenCalledWith('/admin/users', { params: { skip: 2, limit: 10, envelope: true } })
   })
 
   it('updateUser PUT with body', async () => {
     api.put.mockResolvedValueOnce(ok({}))
-    await AdminApiService.updateUser(7, { is_active: false })
+    await AdminApi.updateUser(7, { is_active: false })
     expect(api.put).toHaveBeenCalledWith('/admin/users/7', { is_active: false })
   })
 
   it('bulkUserAction POST with payload', async () => {
     api.post.mockResolvedValueOnce(ok({}))
-    await AdminApiService.bulkUserAction([1,2,3], 'activate')
+    await AdminApi.bulkUserAction([1,2,3], 'activate')
     expect(api.post).toHaveBeenCalledWith('/admin/users/bulk-action', { user_ids: [1,2,3], action: 'activate' })
   })
 })
-
-
