+++ frontend/src/services/__tests__/adminApi.responses.test.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
-import AdminApiService, { SystemStats, UserActivity, SystemMetrics, SecurityAudit, MaintenanceSchedules, LogEntry, AuditEntry, UserWithRoles, DatabaseHealth } from '../adminApi'
+import * as AdminApi from '../admin'
+import type { SystemStats, UserActivity, SystemMetrics, SecurityAudit, MaintenanceSchedules, LogEntry, AuditEntry, UserWithRoles, DatabaseHealth } from '../admin'
 import apiClient from '../api'
 
 vi.mock('../api', () => {
@@ -16,7 +17,7 @@ vi.mock('../api', () => {
 const api = apiClient as unknown as { get: any; post: any; put: any; delete: any }
 const ok = (data: any) => Promise.resolve({ data })
 
-describe('AdminApiService response shapes', () => {
+describe('AdminApi response shapes', () => {
     beforeEach(() => vi.clearAllMocks())
 
     it('returns SystemStats shape', async () => {
@@ -28,7 +29,7 @@ describe('AdminApiService response shapes', () => {
             performance: { avg_file_size_mb: 2.5 },
         }
         api.get.mockResolvedValueOnce(ok(data))
-        const res = await AdminApiService.getSystemStats()
+        const res = await AdminApi.getSystemStats()
         expect(res).toEqual(data)
     })
 
@@ -45,7 +46,7 @@ describe('AdminApiService response shapes', () => {
             },
         ]
         api.get.mockResolvedValueOnce(ok(items))
-        const res = await AdminApiService.getUserActivity(1)
+        const res = await AdminApi.getUserActivity(1)
         expect(res).toEqual(items)
     })
 
@@ -60,7 +61,7 @@ describe('AdminApiService response shapes', () => {
             avg_response_time: 120,
         }
         api.get.mockResolvedValueOnce(ok(metrics))
-        const res = await AdminApiService.getSystemMetrics()
+        const res = await AdminApi.getSystemMetrics()
         expect(res).toEqual(metrics)
     })
 
@@ -73,7 +74,7 @@ describe('AdminApiService response shapes', () => {
             recommendations: ['Enable 2FA'],
         }
         api.get.mockResolvedValueOnce(ok(audit))
-        const res = await AdminApiService.getSecurityAudit()
+        const res = await AdminApi.getSecurityAudit()
         expect(res.recommendations[0]).toBe('Enable 2FA')
     })
 
@@ -92,43 +93,41 @@ describe('AdminApiService response shapes', () => {
             roles: ['admin'],
         }
         api.get.mockResolvedValueOnce(ok({ items: [user], skip: 0, limit: 1, total: 1 }))
-        const env = await AdminApiService.listUsers(0, 1, true) as any
+        const env = await AdminApi.listUsers(0, 1, true) as any
         expect(env.items[0].username).toBe('alice')
 
         api.get.mockResolvedValueOnce(ok([user]))
-        const raw = await AdminApiService.listUsers(0, 1, false) as any
+        const raw = await AdminApi.listUsers(0, 1, false) as any
         expect(raw[0].roles).toContain('admin')
     })
 
     it('returns logs envelope and raw', async () => {
         const log: LogEntry = { timestamp: new Date().toISOString(), level: 'ERROR', message: 'x', module: 'core', user_id: null }
         api.get.mockResolvedValueOnce(ok({ items: [log], skip: 0, limit: 100, total: 1 }))
-        const env = await AdminApiService.getSystemLogs('ERROR', 100, { envelope: true }) as any
+        const env = await AdminApi.getSystemLogs('ERROR', 100, { envelope: true }) as any
         expect(env.items.length).toBe(1)
 
         api.get.mockResolvedValueOnce(ok([log]))
-        const raw = await AdminApiService.getSystemLogs('ERROR', 100) as any
+        const raw = await AdminApi.getSystemLogs('ERROR', 100) as any
         expect(Array.isArray(raw)).toBe(true)
     })
 
     it('returns audit envelope variants', async () => {
         const entry: AuditEntry = { timestamp: new Date().toISOString(), action: 'LOGIN', user_id: 1 }
         api.get.mockResolvedValueOnce(ok({ items: [entry], skip: 0, limit: 100, total: 1 }))
-        const env = await AdminApiService.getAuditLogs(0, 100, undefined, undefined, { envelope: true }) as any
+        const env = await AdminApi.getAuditLogs(0, 100, undefined, undefined, { envelope: true }) as any
         expect(env.items[0].action).toBe('LOGIN')
     })
 
     it('returns health and maintenance schedules', async () => {
         const db: DatabaseHealth = { status: 'ok', connection_pool: {}, table_statistics: {}, performance_metrics: {} }
         api.get.mockResolvedValueOnce(ok(db))
-        const dbRes = await AdminApiService.getDatabaseHealth()
+        const dbRes = await AdminApi.getDatabaseHealth()
         expect(dbRes.status).toBe('ok')
 
         const schedules: MaintenanceSchedules = { items: [{ id: '1', name: 'cleanup', task: 'cleanup', schedule: '* * * * *', enabled: true }] }
         api.get.mockResolvedValueOnce(ok(schedules))
-        const schedRes = await AdminApiService.getMaintenanceSchedules()
+        const schedRes = await AdminApi.getMaintenanceSchedules()
         expect(schedRes.items[0].task).toBe('cleanup')
     })
 })
-
-
