+++ frontend/src/services/admin/index.ts
@@ -1,0 +1,5 @@
+export * from './users';
+export * from './system';
+export * from './logs';
+export * from './maintenance';
+export * from './reports';
